+++ CHANGELOG.md
@@ -588,3 +588,4 @@
 - [2026-04-23 12:22] docs(ui): update readme and API contracts for ui
 - [2026-04-23 13:05] feat(auth): integrate custom telemetry logger in auth
 - [2026-04-23 13:09] feat(navigation): add bi-directional event bus handler for navigation
+- [2026-04-23 13:55] perf(navigation): apply RepaintBoundary on animated navigation widgets